habit-tracker-agent/backend/requirements.txt

fastapi
uvicorn
google-generativeai
google-adk
python-dotenv
